Refactor Spotify class to use less queries
Before, I was always pulling exactly one item for the first page of any
query to their API for a listing. With this, I could take the total,
determine the random index to fetch, and then issue a second request to
fetch the correct item, early exiting if the index was 0.

Now, I'm fetching whatever the default limit of the endpoint is,
determining my index off of a random range between 0 and search total,
early exiting if I have the item from the first response and only
issuing a second request if not.

In this way, for artists/labels with less than 20 albums, I should only
have to issue 1 request to get a random item from them.