Save the last search to the settings
I've noticed this issue where, upon revisiting the app after a bit from
an album view and pressing next, it redirects to a URL missing the
search term. If that happens, I have set it up such that if your session
has saved the latest search term, redirect back to it.