Install GrowthBook for A/B testing & feature flags
I originally set this up for a talk I gave with no intention of merging,
but it ended up being a lot of work that turned out really good, so I'm
deciding to keep it.

[Remixing A/B Testing Talk](https://eligundry.com/talks/remix-ab-testing/)

I have setup the following fun features for GrowthBook:

- Features are loaded server side and hydrated client side such that
  there is not a flash of the base case on initial load.
- User attributes, such as country, language, browser, bot status and
  Google Analytics Client ID are collected and factored into feature
  assignment.
- Experiment assignments are piped to Google Analytics and then to
  BiqQuery such that GrowthBook can analyze it.

Currently, nothing is flagged or experimented. The endpoint responds
fast enough and async from content enough that I'm fine having it
running on each request at the current volume.

Some other features added in this commit include:

- Making Partytown work well with GTM's debugger
- UserProvider refactored to push the user ID to Sentry and GA